+++ core/base/build.gradle.kts
@@ -28,6 +28,8 @@ dependencies {
     api("com.squareup.okhttp3:okhttp:${okHttpVersion}")
     api("net.jodah:failsafe:${jodahFailsafeVersion}")
 
+    implementation(project(":core:policy:policy-engine"))
+
     testImplementation("org.awaitility:awaitility:4.1.1")
     testImplementation("org.junit.jupiter:junit-jupiter-api:${jupiterVersion}")
 }
