+++ core/base/src/main/java/org/eclipse/dataspaceconnector/core/base/agent/ParticipantAgentServiceImpl.java
@@ -11,11 +11,11 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.contract.agent;
+package org.eclipse.dataspaceconnector.core.base.agent;
 
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentService;
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentServiceExtension;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgentService;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgentServiceExtension;
 import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 
 import java.util.ArrayList;
