+++ core/base/src/main/java/org/eclipse/dataspaceconnector/core/base/policy/PolicyContextImpl.java
@@ -11,10 +11,10 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.contract.policy;
+package org.eclipse.dataspaceconnector.core.base.policy;
 
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
-import org.eclipse.dataspaceconnector.spi.contract.policy.PolicyContext;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.policy.PolicyContext;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
