+++ extensions/policy/ids-policy/src/test/java/org/eclipse/dataspaceconnector/ids/policy/PartnerLevelConstraintFunctionTest.java
@@ -2,7 +2,7 @@ package org.eclipse.dataspaceconnector.ids.policy;
 
 import org.eclipse.dataspaceconnector.policy.model.Operator;
 import org.eclipse.dataspaceconnector.policy.model.Permission;
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
