+++ spi/contract-spi/src/main/java/org/eclipse/dataspaceconnector/spi/contract/offer/ContractDefinitionService.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.spi.contract.offer;
 
-import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractDefinition;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
