+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/agent/ParticipantAgentServiceExtension.java
@@ -11,7 +11,7 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.spi.contract.agent;
+package org.eclipse.dataspaceconnector.spi.agent;
 
 import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 import org.jetbrains.annotations.NotNull;
